requirements.txt
discord.py[voice] ~= 1.0.0a